requirements.txt
bidict==0.23.1
certifi==2025.7.9
charset-normalizer==3.4.2
h11==0.16.0
idna==3.10
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.4
simple-websocket==1.1.0
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
